requirements.txt
streamlit
langchain-openai
langchain-core
python-dotenv
gtts
pygame
